Make Xbox 360 builds possible (I don't know how to make DLLs load in Xenia, though), general improvements

+++ base/init_win32.cpp
@@ -5,13 +5,13 @@ BASEAPI void Plat_Init()
 {
 	if (!Base_InitLoader())
 	{
-		Base_Quit((u32)NtCurrentTeb()->LastStatusValue, "Failed to initialize dynamic loader");
+		Base_Quit(NtCurrentTeb()->LastStatusValue, "Failed to initialize dynamic loader");
 	}
 
 	NTSTATUS status = NtQuerySystemInformation(SystemBasicInformation, &g_systemInfo, sizeof(SYSTEM_BASIC_INFORMATION), nullptr);
 	if (!NT_SUCCESS(status))
 	{
-		Base_Quit((u32)status, "Failed to get basic system information");
+		Base_Quit(status, "Failed to get basic system information");
 	}
 }
